重构代码
